// src/app/model/color.type.ts
export type TailwindColor =
  | 'gray'
  | 'red'
  | 'yellow'
  | 'green'
  | 'blue'
  | 'indigo'
  | 'purple'
  | 'pink'
  | 'sky'
  // | 'white'
  | 'black'
  | 'orange'
  | 'teal'
  | 'lime'
  | 'fuchsia'
  | 'black';
//   | 'gray-light'
//   | 'red-light'
//   | 'yellow-light'
//   | 'green-light'
//   | 'blue-light'
//   | 'indigo-light'
//   | 'purple-light'
//   | 'pink-ligh t'
//   | 'white-light'
//   | 'black-light'
//   | 'gray-dark'
//   | 'red-dark'
//   | 'yellow-dark'
//   | 'green-dark'
//   | 'blue-dark'
//   | 'indigo-dark'
//   | 'purple-dark'
//   | 'pink-dark'
//   | 'white-dark'
//   | 'black-dark';